SearchFragment ready

+++ app/src/main/java/com/example/tinkofflab2023/data/remote/response/search/SearchResponse.kt
@@ -13,5 +13,5 @@ data class SearchResponse(
     @SerializedName("personaname")
     val personaname: String,
     @SerializedName("similarity")
-    val similarity: Int
+    val similarity: Double
 )
